+++ lib/helpers/user_helper.dart
@@ -1,5 +1,5 @@
 import 'package:boxify/app_core.dart';
-import 'package:boxify/widgets/must_login_dialog.dart';
+import 'package:boxify/widgets/login_required_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
@@ -10,13 +10,14 @@ class UserHelper {
   /// Returns true if the user is logged in, or prompts the user to log in if not (and returns false).
   /// The action name passed as a parameter is displayed in the dialog (e.g. 'rate tracks', 'create playlists', etc..).
   static bool isLoggedInOrReroute(
-      UserState userState, BuildContext context, String action, IconData icon) {
+      UserState userState, BuildContext context, String action, IconData icon,
+      {bool useRootNavigator = true}) {
     // Is the user logged in?
     if (userState.user.isAnonymous) {
       // Show 'you must be logged in' dialog
       showDialog(
         context: context,
-        useRootNavigator: false,
+        useRootNavigator: useRootNavigator,
         builder: (BuildContext context) {
           return LoginRequiredDialog(
             icon: icon,
