+++ lib/widgets/login_required_dialog.dart
